+++ src/ARte/core/views.py
@@ -8,9 +8,8 @@ from django.views.decorators.cache import cache_page
 
 from .helpers import handle_upload_image
 from .forms import UploadFileForm, ExhibitForm
-from .models import Exhibit
-from users.models import Artwork, Marker, Object
-
+from .models import Exhibit, Artwork, Marker, Object
+ 
 @cache_page(60 * 60)
 def service_worker(request):
     return render(request, 'core/sw.js',
