+++ src/ARte/users/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
-from users.models import Artwork, Profile, Object
+from users.models import Profile
+from core.models import Artwork, Object
 
 admin.site.register(Artwork)
 admin.site.register(Profile)
