+++ src/ARte/users/views.py
@@ -20,8 +20,7 @@ from django.views.decorators.cache import cache_page
 
 
 from .forms import SignupForm, RecoverPasswordCodeForm, RecoverPasswordForm, UploadMarkerForm, UploadObjectForm, ArtworkForm, ExhibitForm, ProfileForm, PasswordChangeForm
-from .models import Marker, Object, Artwork, Profile
-from core.models import Exhibit
+from core.models import Exhibit, Marker, Object, Artwork
 from core.helpers import *
 
 def signup(request):
